basic styling added with header
