+++ src/components/Announcements/announcements.module.css
@@ -1,0 +1,23 @@
+.index {
+  width: 50%;
+  margin: 0 auto;
+  text-align: center;
+  display: block;
+  padding: 30px;
+  
+}
+@media(min-width: 700px) {
+  .index {
+    display: inline-block;
+  }
+}
+
+.index img {
+  width: 100%;
+}
+
+.index h2 {
+  font-size: 3rem;
+  color: rgba(255,255,255, .95);
+  letter-spacing: 5px;
+}
