+++ src/components/Header/Logo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import logo from '../../static/the_lunch_podcast_logo.png';
-import styles from '../../styles/header.module.css';
+import styles from './header.module.css';
 
 
 export default function Logo () {
