+++ src/components/Header/header.module.css
@@ -1,0 +1,25 @@
+.header {
+    padding: 20px;
+}
+
+.header .logo {
+    width: 20%;
+    display: inline-block;
+}
+.header nav {
+    float: right;
+}
+.header nav li {
+    list-style-type: none;
+}
+
+.header li a {
+    font-size: 2.4rem;
+    text-decoration: none;
+    color: rgba(255,255,255, .9);
+    border: 5px solid rgba(255,255,255, .9);
+    padding: 10px;
+}
+.header li a:hover {
+    text-decoration: underline;
+}
