+++ src/components/home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import TwitterFeed from "./twitter_feed";
+import TwitterFeed from "./TwitterFeed";
 import Announcements from './Announcements';
 
 export default function() {
