+++ src/styles/App.css
@@ -24,24 +24,6 @@ html {
   color: #61dafb;
 }
 
-.twitter-feed {
-  max-width: 300px;
-  width: 80%;
-  min-width: 180px;
-  position: relative;
-  margin: 0 auto;
-  max-height: 500px;
-  overflow: scroll;
-}
-@media(min-width: 700px) {
-  .twitter-feed {
-    float: right;
-  }
-}
-.twitter-timeline {
-  width: 100%;
-}
-
 a {
   /* max-width: 200px; */
   position: relative;
