--- src/styles/header.module.css
@@ -1,9 +1,0 @@
-.header {
-    background: red;
-}
-
-.logo {
-    width: 10%;
-    display: inline-block;
-    margin: 20px;
-}
